+++ platformio/commands/platform.py
@@ -299,14 +299,20 @@ def platform_show(platform, json_output):  # pylint: disable=too-many-branches
 @click.option("--with-package", multiple=True)
 @click.option("--without-package", multiple=True)
 @click.option("--skip-default-package", is_flag=True)
+@click.option("--with-all-packages", is_flag=True)
 @click.option(
     "-f",
     "--force",
     is_flag=True,
     help="Reinstall/redownload dev/platform and its packages if exist",
 )
-def platform_install(
-    platforms, with_package, without_package, skip_default_package, force
+def platform_install(  # pylint: disable=too-many-arguments
+    platforms,
+    with_package,
+    without_package,
+    skip_default_package,
+    with_all_packages,
+    force,
 ):
     pm = PlatformManager()
     for platform in platforms:
@@ -315,6 +321,7 @@ def platform_install(
             with_packages=with_package,
             without_packages=without_package,
             skip_default_package=skip_default_package,
+            with_all_packages=with_all_packages,
             force=force,
         ):
             click.secho(